refactor(payment): PAYPAL-3602 updated imports for paypal commerce analytic tracker
