+++ packages/core/src/analytics/paypal-commerce-analytic-tracker/index.ts
@@ -1,0 +1,2 @@
+export { default as createPayPalCommerceAnalyticTracker } from './create-paypal-commerce-analytic-tracker';
+export { default as PayPalCommerceAnalyticTrackerService } from './paypal-commerce-analytic-tracker-service';
