--- packages/core/src/analytics/paypal-commerce-connect-tracker/index.ts
@@ -1,4 +1,0 @@
-// TODO: update createPayPalCommerceConnectTracker with createPayPalCommerceAnalyticTracker
-export { default as createPayPalCommerceConnectTracker } from './create-paypal-commerce-analytic-tracker';
-// TODO: update PayPalCommerceConnectTrackerService with PayPalCommerceAnalyticTrackerService
-export { default as PayPalCommerceConnectTrackerService } from './paypal-commerce-analytic-tracker-service';
